+++ src/element/index.ts
@@ -10,3 +10,4 @@ export { hitTest } from "./collision";
 export { resizeTest } from "./resizeTest";
 export { isTextElement } from "./typeChecks";
 export { textWysiwyg } from "./textWysiwyg";
+export { redrawTextBoundingBox } from "./textElement";
